feat: implement base validator builder
