+++ Cargo.toml
@@ -160,10 +160,8 @@ base-builder-metering = { path = "crates/builder/metering" }
 # Client
 base-client-cli = { path = "crates/client/cli" }
 base-metering = { path = "crates/client/metering" }
-base-client-engine = { path = "crates/client/engine" }
 base-flashblocks = { path = "crates/client/flashblocks" }
 base-txpool-tracing = { path = "crates/client/txpool-tracing" }
-base-engine-tree = { path = "crates/client/engine-tree" }
 
 # Consensus
 base-consensus-rpc = { path = "crates/consensus/rpc" }
@@ -197,6 +195,7 @@ base-execution-txpool = { path = "crates/execution/txpool" }
 base-node-core = { path = "crates/execution/node" }
 base-node-runner = { path = "crates/execution/runner" }
 base-txpool-rpc = { path = "crates/execution/txpool-rpc" }
+base-engine-tree = { path = "crates/execution/engine-tree" }
 
 # Infra
 based = { path = "crates/infra/based" }
