--- crates/client/engine/Cargo.toml
@@ -1,73 +1,0 @@
-[package]
-name = "base-client-engine"
-description = "Engine validator for Base Node"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# workspace
-base-flashblocks.workspace = true
-
-# reth
-reth-evm.workspace = true
-reth-rpc.workspace = true
-reth-trie.workspace = true
-reth-revm.workspace = true
-reth-errors.workspace = true
-reth-rpc-api.workspace = true
-reth-tracing.workspace = true
-reth-provider.workspace = true
-reth-chainspec.workspace = true
-reth-consensus.workspace = true
-reth-node-core.workspace = true
-reth-node-api.workspace = true
-reth-rpc-builder.workspace = true
-reth-tokio-util.workspace = true
-reth-chain-state.workspace = true
-reth-engine-tree.workspace = true
-reth-node-builder.workspace = true
-reth-rpc-eth-types.workspace = true
-reth-trie-parallel.workspace = true
-reth-payload-builder.workspace = true
-reth-rpc-engine-api.workspace = true
-reth-primitives-traits.workspace = true
-reth-engine-primitives.workspace = true
-reth-payload-primitives.workspace = true
-reth-tasks.workspace = true
-
-# revm
-revm-primitives.workspace = true
-
-# alloy
-alloy-evm.workspace = true
-alloy-eips = { workspace = true, features = ["std"] }
-alloy-rpc-types.workspace = true
-alloy-consensus = { workspace = true, features = ["std"] }
-alloy-primitives.workspace = true
-alloy-rpc-types-engine = { workspace = true, features = ["std"] }
-
-# rpc
-jsonrpsee.workspace = true
-
-# misc
-eyre.workspace = true
-rayon.workspace = true
-tracing = { workspace = true, features = ["std"] }
-derive_more.workspace = true
-base-node-core.workspace = true
-base-alloy-evm.workspace = true
-
-base-execution-chainspec.workspace = true
-base-execution-primitives.workspace = true
-base-alloy-consensus.workspace = true
-base-alloy-rpc-types-engine.workspace = true
-base-revm.workspace = true
-base-engine-tree.workspace = true
-revm.workspace = true
