+++ crates/client/flashblocks-node/Cargo.toml
@@ -41,6 +41,9 @@ test-utils = [
 # workspace
 base-flashblocks.workspace = true
 base-node-runner.workspace = true
+base-node-core.workspace = true
+base-engine-tree.workspace = true
+
 # reth
 reth-chain-state.workspace = true
 reth-provider = { workspace = true, optional = true }
