+++ crates/execution/engine-tree/Cargo.toml
@@ -22,14 +22,17 @@ reth-payload-primitives.workspace = true
 reth-primitives-traits.workspace = true
 reth-provider.workspace = true
 reth-revm.workspace = true
+reth-node-api.workspace = true
+reth-node-builder.workspace = true
 reth-trie.workspace = true
 reth-trie-parallel.workspace = true
-reth-trie-db.workspace = true
 base-alloy-consensus.workspace = true
 base-revm.workspace = true
 base-execution-primitives.workspace = true
 base-execution-evm.workspace = true
 base-execution-chainspec.workspace = true
+base-flashblocks.workspace = true
+base-node-core.workspace = true
 reth-tasks.workspace = true
 
 revm.workspace = true
@@ -48,3 +51,4 @@ tokio.workspace = true
 tracing.workspace = true
 derive_more.workspace = true
 crossbeam-channel.workspace = true
+eyre.workspace = true
