+++ crates/execution/runner/Cargo.toml
@@ -50,6 +50,7 @@ test-utils = [
 # workspace
 base-cli-utils.workspace = true
 base-test-utils = { workspace = true, optional = true }
+base-engine-tree.workspace = true
 
 # reth
 reth-evm.workspace = true
